Move to utils, and make ModelContainer do it too
