+++ jwst/datamodels/utils/__init__.py
@@ -1,1 +1,5 @@
 """Utilities for manipulation of datamodels."""
+
+from jwst.datamodels.utils.group_id import attrs_to_group_id
+
+__all__ = ["attrs_to_group_id"]
